--- __tests__/fixtures/TIT-1-1.js
@@ -1,3 +1,0 @@
-export const verseObjects = [{ "type": "text", "text": " " }, { "text": "Ἰούδας", "tag": "w", "type": "word", "lemma": "Ἰούδας", "strong": "G24550", "morph": "Gr,N,,,,,NMS,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": ", " }, { "text": "Ἰησοῦ", "tag": "w", "type": "word", "lemma": "Ἰησοῦς", "strong": "G24240", "morph": "Gr,N,,,,,GMS,", "occurrence": 1, "occurrences": 2 }, { "type": "text", "text": " " }, { "text": "Χριστοῦ", "tag": "w", "type": "word", "lemma": "χριστός", "strong": "G55470", "morph": "Gr,N,,,,,GMS,", "tw": "rc://*/tw/dict/bible/kt/christ", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "type": "text", "text": " " }, { "text": "δοῦλος", "tag": "w", "type": "word", "lemma": "δοῦλος", "strong": "G14010", "morph": "Gr,N,,,,,NMS,", "tw": "rc://*/tw/dict/bible/other/servant", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": ", " }, { "text": "ἀδελφὸς", "tag": "w", "type": "word", "lemma": "ἀδελφός", "strong": "G00800", "morph": "Gr,N,,,,,NMS,", "tw": "rc://*/tw/dict/bible/kt/brother", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "δὲ", "tag": "w", "type": "word", "lemma": "δέ", "strong": "G11610", "morph": "Gr,CC,,,,,,,,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "Ἰακώβου", "tag": "w", "type": "word", "lemma": "Ἰάκωβος", "strong": "G23850", "morph": "Gr,N,,,,,GMS,", "tw": "rc://*/tw/dict/bible/names/jamesbrotherofjesus", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": "; " }, { "text": "τοῖς", "tag": "w", "type": "word", "lemma": "ὁ", "strong": "G35880", "morph": "Gr,RD,,,,DMP,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "ἐν", "tag": "w", "type": "word", "lemma": "ἐν", "strong": "G17220", "morph": "Gr,P,,,,,D,,,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "Θεῷ", "tag": "w", "type": "word", "lemma": "θεός", "strong": "G23160", "morph": "Gr,N,,,,,DMS,", "tw": "rc://*/tw/dict/bible/kt/god", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "Πατρὶ", "tag": "w", "type": "word", "lemma": "πατήρ", "strong": "G39620", "morph": "Gr,N,,,,,DMS,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "type": "text", "text": " " }, { "text": "ἠγαπημένοις", "tag": "w", "type": "word", "lemma": "ἀγαπάω", "strong": "G00250", "morph": "Gr,V,PEP,DMP,", "tw": "rc://*/tw/dict/bible/kt/love", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": ", " }, { "text": "καὶ", "tag": "w", "type": "word", "lemma": "καί", "strong": "G25320", "morph": "Gr,CC,,,,,,,,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "Ἰησοῦ", "tag": "w", "type": "word", "lemma": "Ἰησοῦς", "strong": "G24240", "morph": "Gr,N,,,,,DMS,", "occurrence": 2, "occurrences": 2 }, { "type": "text", "text": " " }, { "text": "Χριστῷ", "tag": "w", "type": "word", "lemma": "χριστός", "strong": "G55470", "morph": "Gr,N,,,,,DMS,", "tw": "rc://*/tw/dict/bible/kt/christ", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "type": "text", "text": " " }, { "text": "τετηρημένοις", "tag": "w", "type": "word", "lemma": "τηρέω", "strong": "G50830", "morph": "Gr,V,PEP,DMP,", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": " " }, { "text": "κλητοῖς", "tag": "w", "type": "word", "lemma": "κλητός", "strong": "G28220", "morph": "Gr,NS,,,,DMP,", "tw": "rc://*/tw/dict/bible/kt/call", "occurrence": 1, "occurrences": 1 }, { "type": "text", "text": ":" }];
-export const quote = `Ἰούδας, Ἰησοῦ Χριστοῦ  δοῦλος,`;
-export const occurrence = 1;
