+++ __tests__/fixtures/highlighting/tit-1-1.js
@@ -1,0 +1,6 @@
+module.exports = {
+  book: 'tit',
+  chapter: 1,
+  verse: 1,
+  quote: `τῆς κατ’ εὐσέβειαν`
+}
