+++ __tests__/fixtures/highlighting/tit-1-2.js
@@ -1,0 +1,6 @@
+module.exports = {
+  book: 'tit',
+  chapter: 1,
+  verse: 2,
+  quote: `ἐπ’ ἐλπίδι ζωῆς αἰωνίου`
+}
