+++ __tests__/fixtures/highlighting/tit-1-3.js
@@ -1,0 +1,6 @@
+module.exports = {
+  book: 'tit',
+  chapter: 1,
+  verse: 3,
+  quote: `τοῦ Σωτῆρος ἡμῶν, Θεοῦ`
+}
